Run the standalone-Flows guard on any pipecat import
The coexistence guard lived in pipecat.flows.__init__, so it only ran when
something imported pipecat.flows. But an app still on the standalone package
does `from pipecat_flows import ...`, which pulls in core pipecat (and its
__init__) but never pipecat.flows — so the guard never fired, and the most
likely misconfiguration (running an old Flows app against a Flows-bearing
Pipecat) went unwarned.

Move the check to the top-level pipecat/__init__.py, which runs on any pipecat
import. Gate it like the version banner so it stays quiet for the pipecat/pc
CLI. Reword the changelog accordingly.
