+++ src/pipecat/__init__.py
@@ -4,6 +4,7 @@
 # SPDX-License-Identifier: BSD 2-Clause License
 #
 
+import importlib.util
 import os
 import sys
 from importlib.metadata import version as lib_version
@@ -24,8 +25,35 @@ def _should_log_version_banner() -> bool:
     return program not in ("pipecat", "pc")
 
 
+def _warn_if_standalone_flows_installed() -> None:
+    """Flag the deprecated standalone ``pipecat-ai-flows`` package if it is also installed.
+
+    Pipecat Flows now ships inside ``pipecat-ai`` as ``pipecat.flows``. Older
+    ``pipecat-ai-flows`` releases allow ``pipecat-ai<2``, so they can end up
+    installed next to a Pipecat that already includes Flows — a redundant, easily
+    confused setup. The check lives here, in the top-level package init that runs
+    on any ``pipecat`` import, rather than in ``pipecat.flows`` — so it also fires
+    for apps still importing the standalone ``pipecat_flows`` (which pulls in core
+    Pipecat but never ``pipecat.flows``). Detection uses ``find_spec`` to avoid
+    importing the standalone package.
+    """
+    try:
+        installed = importlib.util.find_spec("pipecat_flows") is not None
+    except (ImportError, ValueError):
+        installed = False
+    if installed:
+        logger.error(
+            "The separate `pipecat-ai-flows` package is installed alongside a version "
+            "of Pipecat that already includes Pipecat Flows as `pipecat.flows`. You do "
+            "not need both — uninstall `pipecat-ai-flows` and import Flows from "
+            "`pipecat.flows`."
+        )
+
+
 if _should_log_version_banner():
     logger.info(f"ᓚᘏᗢ Pipecat {__version__} (Python {sys.version}) ᓚᘏᗢ")
+    # Gated like the banner: skip the redundant-package check for the pipecat/pc CLI.
+    _warn_if_standalone_flows_installed()
 
 
 def version() -> str:
